Implemeted String#unpack('m'). Moved pack/unpack format string parser to a separate file.
